added parse_url function
function that parses the passed url. If the url is that of a supported
website, it continues the parsing, otherwise it raises a TypeError. If
the url is that of a playlist or channel, it returns a url list of it's
contents, otherwise returns the same url given as an argument.